+++ proposals/description/fip_54.ts
@@ -15,7 +15,7 @@ const permanently_revoke_burner: ProposalDescription = {
   Replace the core reference in the FEI token to a “Restricted Permissions” which only allows for minting and pausing. 
   This would permanently lock the contract’s ability to burn from any address. It preserves the ability for a user or contract to burn its own FEI.
 
-  Code: https://github.com/fei-protocol/fei-protocol-core/blob/feat/v2/base/contracts/core/RestrictedPermissions.sol
+  Code: https://github.com/fei-protocol/fei-protocol-core/pull/352
   Discussion: https://tribe.fei.money/t/fip-54-permanently-deprecate-burner/3743 
 `
 };
